refactor(LineAcceptanceTest): 테스트코드 fixture 클래스분리
- setUp 분리, 상속
- fixture 메소드 추가;지하철_노선_생성
