+++ src/test/java/nextstep/subway/line/LineAcceptanceTest.java
@@ -9,11 +9,9 @@ import java.util.Arrays;
 import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
-import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 import org.springframework.boot.test.context.SpringBootTest;
-import org.springframework.boot.web.server.LocalServerPort;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.MediaType;
 import org.springframework.test.annotation.DirtiesContext;
@@ -21,15 +19,7 @@ import org.springframework.test.annotation.DirtiesContext;
 @DisplayName("지하철 노선 관련 기능")
 @DirtiesContext(classMode = DirtiesContext.ClassMode.BEFORE_EACH_TEST_METHOD)
 @SpringBootTest(webEnvironment = SpringBootTest.WebEnvironment.RANDOM_PORT)
-public class LineAcceptanceTest {
-
-    @LocalServerPort
-    int port;
-
-    @BeforeEach
-    public void setUp() {
-        RestAssured.port = port;
-    }
+public class LineAcceptanceTest extends LineAcceptanceTestFixture{
 
     /**
      * When 지하철 노선을 생성하면
@@ -39,14 +29,7 @@ public class LineAcceptanceTest {
     @Test
     void 지하철_노선_생성_확인() {
         //When
-        Map<String, String> params = new HashMap<>();
-        params.put("name", "2호선");
-        ExtractableResponse<Response> response = RestAssured.given().log().all()
-                .body(params)
-                .contentType(MediaType.APPLICATION_JSON_VALUE)
-                .when().post("/lines")
-                .then().log().all()
-                .extract();
+        ExtractableResponse<Response> response = 지하철_노선_생성("2호선");
 
         //Then
         assertThat(response.statusCode()).isEqualTo(HttpStatus.CREATED.value());
@@ -70,17 +53,7 @@ public class LineAcceptanceTest {
     void 지하철_노선_목록_조회_확인() {
         //Given
         List<String> createLines = Arrays.asList("2호선", "3호선");
-        createLines.forEach(l -> {
-            Map<String, String> params = new HashMap<>();
-            params.put("name", l);
-
-            RestAssured.given().log().all()
-                    .body(params)
-                    .contentType(MediaType.APPLICATION_JSON_VALUE)
-                    .when().post("/lines")
-                    .then().log().all()
-                    .extract();
-        });
+        createLines.forEach(LineAcceptanceTestFixture::지하철_노선_생성);
 
         //When
         ExtractableResponse<Response> response = RestAssured.given().log().all()
@@ -106,15 +79,7 @@ public class LineAcceptanceTest {
     @Test
     void 지하철_노선_조회_확인() {
         //Given
-        Map<String, String> params = new HashMap<>();
-        params.put("name", "2호선");
-
-        int lineId = RestAssured.given().log().all()
-                .body(params)
-                .contentType(MediaType.APPLICATION_JSON_VALUE)
-                .when().post("/lines")
-                .then().log().all()
-                .extract().jsonPath().get("id");
+        int lineId = 지하철_노선_생성("2호선").jsonPath().get("id");
 
         //When
         ExtractableResponse<Response> response = RestAssured.given().log().all()
@@ -140,15 +105,7 @@ public class LineAcceptanceTest {
     @Test
     void 지하철_노선_수정_확인() {
         //Given
-        Map<String, String> params = new HashMap<>();
-        params.put("name", "2호선");
-
-        int lineId = RestAssured.given().log().all()
-                .body(params)
-                .contentType(MediaType.APPLICATION_JSON_VALUE)
-                .when().post("/lines")
-                .then().log().all()
-                .extract().jsonPath().get("id");
+        int lineId = 지하철_노선_생성("2호선").jsonPath().get("id");
 
         //When
         Map<String, String> patchParams = new HashMap<>();
@@ -182,15 +139,7 @@ public class LineAcceptanceTest {
     @Test
     void 지하철_노선_삭제_확인() {
         //Given
-        Map<String, String> params = new HashMap<>();
-        params.put("name", "2호선");
-
-        int lineId = RestAssured.given().log().all()
-                .body(params)
-                .contentType(MediaType.APPLICATION_JSON_VALUE)
-                .when().post("/lines")
-                .then().log().all()
-                .extract().jsonPath().get("id");
+        int lineId = 지하철_노선_생성("2호선").jsonPath().get("id");
 
         //When
         ExtractableResponse<Response> response =
